Move project creation logic from CLI to lib
